refactor: 도메인별 Service 재정의

+++ src/main/kotlin/io/hhplus/concert/domain/queue/QueueService.kt
@@ -1,31 +1,35 @@
 package io.hhplus.concert.domain.queue
 
+import io.hhplus.concert.core.exception.BizError
 import io.hhplus.concert.core.exception.BizException
-import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 
 @Service
 class QueueService(private val repo: QueueRepository) {
 
-    fun create(userId: Long, scheduleId: Long) = repo.save(userId, scheduleId)
-        .run { QueueInfo(this) }
+    fun create(userId: Long, scheduleId: Long): Queue {
+        if (repo.findByUserId(userId) != null)
+            throw BizException(BizError.Queue.DUPLICATED)
 
-    fun getByUserId(userId: Long) = repo.findByUserId(userId)
-        ?.run { QueueInfo(this) }
-        ?: throw BizException(HttpStatus.NOT_FOUND, NOT_FOUND_MESSAGE)
-
-    fun getAllByScheduleId(scheduleId: Long) = repo.findAllByScheduleId(scheduleId)
-        .map { QueueInfo(it) }
-
-    fun delete(queueId: Long) {
-        val queue = repo.findById(queueId)
-            ?: throw BizException(HttpStatus.NOT_FOUND, NOT_FOUND_MESSAGE)
-
-        repo.delete(queue)
+        val queue = Queue(userId, scheduleId)
+        return repo.save(queue)
     }
 
-    companion object {
-        private const val NOT_FOUND_MESSAGE = "존재하지 않는 대기열입니다."
+    fun get(id: Long) = repo.findById(id)
+        ?: throw BizException(BizError.Queue.NOT_FOUND)
+
+    fun pass(id: Long): Queue {
+        val queue = repo.findById(id)
+            ?: throw BizException(BizError.Queue.NOT_FOUND)
+
+        queue.pass()
+
+        return queue
+    }
+
+    fun delete(id: Long) {
+        val queue = repo.findById(id)
+        if (queue != null) repo.delete(queue)
     }
 
 }
